type conversion

+++ eeg_rpsd.m
@@ -15,7 +15,7 @@ end
 ncomp = size(EEG.icaweights, 1);
 n_points = floor(min(EEG.pnts, EEG.srate));
 try
-    window = windows('hamming', n_points, 0.54)';
+    window = windows('hamming', double(n_points), 0.54)';
 catch
     lasterr
     error('The windows function is in the firfilt plugin, make sure it is in the path')
